sentence-similarity/requirements.txt
sentence_transformer
